docs: address review comments for dynamic formats list
